Move main() from server to application
